Teardown fixes (#81)
* Delete only explicitly created NICs on teardown

Also only EIPs of the target instance

* Delete NICs before EIPs also in teardown_region()

As EIPs depend on NICs

* Don't even try to delete backup if we don't have a bucket set

+++ pg_spot_operator/operator.py
@@ -30,7 +30,7 @@ from pg_spot_operator.cloud_impl.aws_vm import (
     ensure_spot_vm,
     get_addresses,
     get_all_active_operator_instances_in_region,
-    get_network_interfaces,
+    get_non_self_terminating_network_interfaces,
     get_operator_volumes_in_region,
     release_address_by_allocation_id_in_region,
     terminate_instances_in_region,
@@ -737,31 +737,35 @@ def destroy_instance(
             logger.info("Deleting VolumeId %s (%s GB) ...", vol_id, size)
             delete_volume_in_region(m.region, vol_id)
 
+    logger.info("Looking for explicit NICs to delete ....")
+    nic_ids = get_non_self_terminating_network_interfaces(
+        m.region, m.instance_name
+    )
+    logger.info("NICs found: %s", nic_ids)
+    if not dry_run and nic_ids:
+        if not vol_ids_and_sizes:
+            logger.info("OK. Sleeping 60s before deleting NICs ...")
+            time.sleep(60)
+        for nic_id in nic_ids:
+            logger.info("Deleting NIC %s ...", nic_id)
+            delete_network_interface(m.region, nic_id)
+
     logger.info("Looking for Elastic IPs to delete ....")
-    eip_alloc_ids = get_addresses(m.region)
+    eip_alloc_ids = get_addresses(m.region, m.instance_name)
     logger.info("Elastic IP Addresses found: %s", eip_alloc_ids)
     if not dry_run and eip_alloc_ids:
+        if nic_ids:
+            logger.info("Sleeping 60s before deleting EIPs ...")
+            time.sleep(60)
         for alloc_id in eip_alloc_ids:
             logger.info("Releasing Address with AllocationId %s ...", alloc_id)
             release_address_by_allocation_id_in_region(m.region, alloc_id)
-
-    if not m.floating_ips:
-        logger.info("Looking for NICs to delete ....")
-        nic_ids = get_network_interfaces(m.region, m.instance_name)
-        logger.info("NICs found: %s", nic_ids)
-        if not dry_run and nic_ids:
-            if not vol_ids_and_sizes:
-                logger.info("OK. Sleeping 60s before deleting NICs ...")
-                time.sleep(60)
-            for nic_id in nic_ids:
-                logger.info("Deleting NIC %s ...", nic_id)
-                delete_network_interface(m.region, nic_id)
 
     logger.info(
         "OK - cloud resources for instance %s cleaned-up", m.instance_name
     )
 
-    if m.backup.destroy_backups:
+    if m.backup.destroy_backups and m.backup.s3_bucket:
         destroy_backups_if_any(m)
 
     cmdb.finalize_destroy_instance(m)
@@ -813,28 +817,32 @@ def teardown_region(
                     )
                     delete_volume_in_region(region, vol_id)
 
-            logger.info("Looking for EIPs to delete ....")
-            elastic_address_alloc_ids = get_addresses(region)
-            logger.info(
-                "Elastic Addresses found: %s", elastic_address_alloc_ids
-            )
-            if not dry_run and elastic_address_alloc_ids:
-                for alloc_id in elastic_address_alloc_ids:
-                    logger.info(
-                        "Releasein Address wiht AllocationId %s ...", alloc_id
-                    )
-                    release_address_by_allocation_id_in_region(
-                        region, alloc_id
-                    )
-
-            logger.info("Looking for NICs to delete ....")
-            nic_ids = get_network_interfaces(region)
+            logger.info("Looking for explicit NICs to delete ....")
+            nic_ids = get_non_self_terminating_network_interfaces(region)
             logger.info("NICs found: %s", nic_ids)
             if not dry_run and nic_ids:
                 for nic_id in nic_ids:
                     logger.info("Deleting NIC %s ...", nic_id)
                     delete_network_interface(region, nic_id)
             logger.info("Cleanup loop completed")
+
+            logger.info("Looking for EIPs to delete ....")
+            elastic_address_alloc_ids = get_addresses(region)
+            logger.info(
+                "Elastic Addresses found: %s", elastic_address_alloc_ids
+            )
+            if not dry_run and elastic_address_alloc_ids:
+                if nic_ids:
+                    logger.info("Sleeping 60s before deleting EIPs ...")
+                    time.sleep(60)
+                for alloc_id in elastic_address_alloc_ids:
+                    logger.info(
+                        "Releasing Elastic Address with AllocationId %s ...",
+                        alloc_id,
+                    )
+                    release_address_by_allocation_id_in_region(
+                        region, alloc_id
+                    )
             break
         except Exception:
             logger.exception(f"Failed to complete cleanup loop {i}")
